requirements.txt
# Etapa 1: Buscador por similitud
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
pillow>=10.0.0
faiss-cpu>=1.7.4
gradio>=4.0.0
scikit-learn>=1.3.0

# Etapa 2: Entrenamiento de modelos
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Etapa 3: Detección con YOLO
ultralytics>=8.0.0
opencv-python>=4.8.0

# Etapa 4: Optimización y anotación
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0
pycocotools>=2.0.6

# Utilidades generales
jupyter>=1.0.0
ipykernel>=6.25.0


